projetos/infraestrutura_bigquery/scripts/requirements.txt
pandas
google-cloud-bigquery
pyarrow
pandas-gbq
tdqm
google-cloud-storage